Move New Credentials to separate component for Add Tenant
